CASMCMS-6651: remove note about git clone on PIT node
CASMCMS-6649 added a note about cloning the configuration repo on
the PIT node, but customer installs (and internal installs shortly)
of the UAN software will never happen on the PIT node. The documentation
shows the prompts for the commands using `ncn-m001`, which is a master
node, not a pit node.