CASMCMS-6630: hard code UAN repo names
We don't have a good way to version repository names that are
referenced in the image recipe in the SCMS/uan repository.
The version comes from this repository long after the recipe is
wrapped up and placed in a docker image.

Lacking a clearer solution for 1.4, we have decided to hard code
the RPM repository with the eventual UAN version (2.0.0) that is
expected to be released in the Shasta 1.4 timeframe.