CASMCMS-6697: Improve documentation regarding UAN partition creation
Improve the documentation concerning debugging a failure to create a
partition on the UAN node for either swap or scratch space. This
failure occurs when a pre-existing partition prevents the Ansible
play from creating the required, desired partitions.

This is a sub-set of the work described in CASMCMS-6732, which is
to add advanced debugging sections to this document.