CASMCMS-6748: Add UAN preflight check
This goss test checks whether the System Layout Service (SLS) and the
Hardware State Manager (HSM) agree on the number of UAN nodes in the
system. If there is a disagreement, then the test fails and prompts the
user to run a script to determine which service SLS or HSM is missing the
UAN node(s).